+++ aol/lakes/views.py
@@ -1,45 +1,41 @@
 import os.path
 
-from django.utils.decorators import method_decorator
-from django.views.decorators.cache import cache_page
-from django.http import HttpResponse
+from django.views.generic import View
+from django.http import Http404
 from django.conf import settings
 
+from django_sendfile import sendfile
 from rest_framework import generics
 
-from aol.documents.models import Document
-from aol.photos.models import Photo
-
-from . import models, serializers
+from aol.lakes.models import Lake
+from aol.lakes.serializers import LakeDetailSerializer
 
 
-class LakeListView(generics.ListAPIView):
+class LakeListView(View):
     """
     TBD
     """
-    serializer_class = serializers.LakeIndexSerializer
-
-    def get_queryset(self):
+    def get(self, request, *args, **kwargs):
         """
         TBD
         """
-        status = self.request.query_params.get('status', 'major')
-        if status == 'major':
-            return models.Lake.active.major()
-        return models.Lake.active.minor()
+        status = request.GET.get('status', None)
+        render_format = kwargs.get('format', None)
 
-    # @method_decorator(cache_page(60 * 60))
-    def list(self, request, *args, **kwargs):
-        return super().list(request, *args, **kwargs)
+        if status is None:
+            raise Http404
+
+        cached_name = 'lakes-{}.json'.format(status)
+        path = os.path.join(settings.MEDIA_ROOT, cached_name)
+        if not os.path.exists(path):
+            raise Http404
+
+        return sendfile(request, path, mimetype='application/json')
 
 
 class LakeDetailView(generics.RetrieveAPIView):
     """
     TBD
     """
-    queryset = models.Lake.objects.all()
-    serializer_class = serializers.LakeDetailSerializer
-
-    # @method_decorator(cache_page(60 * 60))
-    def retrieve(self, request, *args, **kwargs):
-        return super().retrieve(request, *args, **kwargs)
+    queryset = Lake.objects.all()
+    serializer_class = LakeDetailSerializer